修复超大封包导致无法挡住的问题

+++ util/packet.go
@@ -40,7 +40,7 @@ func RecvLTVPacket(reader io.Reader, maxPacketSize int) (msg interface{}, err er
 	// 用小端格式读取Size
 	size := binary.LittleEndian.Uint16(sizeBuffer)
 
-	if maxPacketSize > 0 && size >= uint16(maxPacketSize) {
+	if maxPacketSize > 0 && int(size) >= maxPacketSize {
 		return nil, ErrMaxPacket
 	}
